Minimised header on scroll and side menu implemented from Botanica project.
The minimised header which replaces the regular header and navigation bar, along with the side menu that pops up when clicking the hamburger icon, have been fitted and added to this project, from my Botanica website. The javascript for the minimised header and the side menu are untouched since then, hence still have the same last updated date.